frontend/requirements.txt
streamlit
requests
pillow
numpy
opencv-python-headless
